+++ src/main/java/blusunrize/immersiveengineering/common/data/Recipes.java
@@ -746,6 +746,13 @@ public class Recipes extends RecipeProvider
 				.addInput(IETags.clay)
 				.setEnergy(3200)
 				.build(out, toRL("mixer/concrete"));
+		MixerRecipeBuilder.builder(IEContent.fluidConcrete, 500)
+				.addFluid(Fluids.WATER, 500)
+				.addInput(new IngredientWithSize(IETags.slag, 2))
+				.addInput(Tags.Items.GRAVEL)
+				.addInput(IETags.clay)
+				.setEnergy(3200)
+				.build(out, toRL("mixer/concrete"));
 	}
 
 	private void mineralMixes(@Nonnull Consumer<IFinishedRecipe> out)
@@ -902,7 +909,7 @@ public class Recipes extends RecipeProvider
 		addCornerStraightMiddle(StoneDecoration.blastbrick, 3,
 				makeIngredient(Tags.Items.INGOTS_NETHER_BRICK),
 				makeIngredient(Tags.Items.INGOTS_BRICK),
-				makeIngredient(Items.BLAZE_POWDER),
+				makeIngredient(Blocks.MAGMA_BLOCK),
 				out);
 		addSandwich(StoneDecoration.hempcrete, 6,
 				makeIngredient(IETags.clay),